# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是後端 API 文件",
        "title": "interchange API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/proxy/{version}/{provider}/chat/completions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "處理聊天生成請求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy-Chat"
                ],
                "summary": "聊天生成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API 版本（例如 v1、v2）",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "模型提供者（例如 openai、azure、gemini）",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "聊天生成請求內容",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_service_chat.ChatPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/interchange_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/interchange_internal_service_chat.ChatResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    }
                }
            }
        },
        "/proxy/{version}/{provider}/embeddings/generations": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "處理嵌入生成請求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy-Embedding"
                ],
                "summary": "嵌入生成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API 版本（例如 v1、v2）",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "模型提供者（例如 openai、azure、gemini）",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "嵌入生成請求內容",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_service_embedding.EmbeddingRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/interchange_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/interchange_internal_service_embedding.EmbeddingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "interchange_internal_pkg_error.Error": {
            "type": "object"
        },
        "interchange_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
...
        "interchange_internal_service_chat.ChatChoice": {
            "type": "object",
            "properties": {
                "extensions": {
                    "description": "可再擴充：如 Gemini 會有自家 output/ext，可加 Extensions",
                    "type": "object",
                    "additionalProperties": true
                },
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "logprobs": {},
                "message": {
                    "$ref": "#/definitions/interchange_internal_service_chat.ChatResponseMessage"
                }
            }
        },
        "interchange_internal_service_chat.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "string (text)、map（image/audio等）、支援多模態"
                },
                "extensions": {
                    "type": "object",
                    "additionalProperties": true
                },
                "function_call": {
                    "description": "可存 openai function call"
                },
                "name": {
                    "description": "optional",
                    "type": "string"
                },
                "role": {
                    "description": "\"system\"/\"user\"/\"assistant\"/...",
                    "type": "string"
                },
                "tool_calls": {
                    "description": "工具調用紀錄"
                }
            }
        },
        "interchange_internal_service_chat.ChatPayload": {
            "type": "object",
            "properties": {
                "audio": {
                    "description": "支援 audio output",
                    "type": "object",
                    "additionalProperties": true
                },
                "extensions": {
                    "description": "Custom platform extension:",
                    "type": "object",
                    "additionalProperties": true
                },
                "frequency_penalty": {
                    "type": "number"
                },
                "function_call": {
                    "description": "Deprecated (string or object)"
                },
                "functions": {
                    "description": "Deprecated (array)"
                },
                "logit_bias": {
                    "description": "token id(string): bias",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "logprobs": {
                    "type": "boolean"
                },
                "max_completion_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_service_chat.ChatMessage"
                    }
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "modalities": {
                    "description": "e.g. [\"text\", \"audio\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "type": "string"
                },
                "n": {
                    "type": "integer"
                },
                "parallel_tool_calls": {
                    "type": "boolean"
                },
                "prediction": {
                    "type": "object",
                    "additionalProperties": true
                },
                "presence_penalty": {
                    "type": "number"
                },
                "prompt_cache_key": {
                    "type": "string"
                },
                "reasoning_effort": {
                    "description": "\"low\"/\"medium\"/\"high\"",
                    "type": "string"
                },
                "response_format": {
                    "description": "e.g. {type:json_schema}",
                    "type": "object",
                    "additionalProperties": true
                },
                "safety_identifier": {
                    "type": "string"
                },
                "seed": {
                    "type": "integer"
                },
                "service_tier": {
                    "description": "\"auto\"/\"default\"/...",
                    "type": "string"
                },
                "stop": {
                    "description": "string or []string or null"
                },
                "store": {
                    "type": "boolean"
                },
                "stream": {
                    "type": "boolean"
                },
                "stream_options": {
                    "type": "object",
                    "additionalProperties": true
                },
                "temperature": {
                    "type": "number"
                },
                "tool_choice": {
                    "description": "string or object"
                },
                "tools": {
                    "description": "array (function/tool desc)"
                },
                "top_logprobs": {
                    "type": "integer"
                },
                "top_p": {
                    "type": "number"
                },
                "user": {
                    "description": "Deprecated, for compat",
                    "type": "string"
                },
                "web_search_options": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "interchange_internal_service_chat.ChatResponseMessage": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "array",
                    "items": {}
                },
                "content": {},
                "extensions": {
                    "description": "支援未來擴充（如 Gemini 多模態等）",
                    "type": "object",
                    "additionalProperties": true
                },
                "refusal": {},
                "role": {
                    "type": "string"
                }
            }
        },
        "interchange_internal_service_chat.ChatResult": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_service_chat.ChatChoice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "extensions": {
                    "description": "供其他平台或自家加欄位",
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "service_tier": {
                    "type": "string"
                },
                "system_fingerprint": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/interchange_internal_service_chat.ChatUsage"
                }
            }
        },
        "interchange_internal_service_chat.ChatUsage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "completion_tokens_details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "extensions": {
                    "type": "object",
                    "additionalProperties": true
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "prompt_tokens_details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "interchange_internal_service_embedding.EmbeddingRequestBody": {
            "type": "object",
            "properties": {
                "dimensions": {
                    "description": "optional",
                    "type": "integer"
                },
                "encoding_format": {
                    "description": "optional: \"float\" or \"base64\"",
                    "type": "string"
                },
                "input": {
                    "description": "string or []string or [][]int"
                },
                "model": {
                    "description": "required",
                    "type": "string"
                },
                "user": {
                    "description": "optional",
                    "type": "string"
                }
            }
        },
        "interchange_internal_service_embedding.EmbeddingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "embedding": {
                                "description": "e.g. 1536 floats for ada-002",
                                "type": "array",
                                "items": {
                                    "type": "number"
                                }
                            },
                            "index": {
                                "type": "integer"
                            },
                            "object": {
                                "type": "string"
                            }
                        }
                    }
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "properties": {
                        "prompt_tokens": {
                            "type": "integer"
                        },
                        "total_tokens": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "請在欄位輸入 \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}